Extend timeout to wait for RPC reply (#7876)
Currently, when processing comm messages, we wait 2s after receiving an
`Idle` before we give up on receiving an RPC reply. Empirically this
doesn't seem to be enough to handle all cases, so this change kicks the
can down the road a bit by waiting 5s.

Mitigates, but doesn't fix,
https://github.com/posit-dev/positron/issues/7755.