Register renv.lock as a file association instead of a language (#7926)
This change fixes an issue that causes all JSON files to show up as
"renv lockfiles". Instead of re-using the `json` ID for these files, we
register `renv.lock` with the JSON file type.

There is a slightly more complicated way to address this that would
allow renv lockfiles to have their own file type with custom
schema/grammar that derives from JSON, but probably not worth it given
that these are not meant to be hand-edited in the first place.

Addresses #7777.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix issue causing all JSON files to have the filetype "renv lockfile"
(#7777)


### QA Notes

Note that after the change, renv lockfiles no longer have a special file
type. They should now just show up as JSON files.