Fix issue resuming from sleep on Windows (#7909)
This change fixes an issue resuming from sleep on Windows. 

The underlying issue was caused by the logic that automatically shuts
down the supervisor when it has become a zombie (i.e. it is not
connected to anything and has been idle a long time). When waking after
sleep, the interval that checks for this state ticks many times to catch
up to clock time, at the same time as the supervisor is processing a
bunch of network issues caused by Windows shutting down its network
stack for sleep. This can cause the supervisor to shut down prematurely.

The fix (which was made over on the supervisor side) is to change the
way the interval timer works so that missed ticks (while the computer is
asleep) are dropped. In this way, we only count idle time while the
computer is awake instead of clock time.

Addresses https://github.com/posit-dev/positron/issues/6660.
Addresses https://github.com/posit-dev/positron/issues/6772.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix issue causing sessions to fail to wake from sleep on Windows
(#6660).


### QA Notes

I was not able to reproduce exactly the behavior observed by @andrie in
#6660, so am only 75% confident this will fix it; there is a chance that
more work is needed here to reconnect ZeroMQ sockets if Windows
disconnects them due to shutting down the network stack during sleep.
Empirically, though, resuming from sleep is successful for me after this
change.