Assistant: Synchronize the current model and provider (#7954)
This PR improves synchronization of the current selected language model
and provider. It also removes the "All Models" option and instead
automatically selects a provider when the first one is registered.
Addresses #7734, #7938.

### QA Notes

1. When there are no registered models, the provider dropdown should be
an "Add Model Provider..." button.
2. When signing in to the first provider, it should get automatically
selected.
3. When signing out of the selected provider, an alternative provider
should be selected, or if there is none the "Add Model Provider..."
button should be displayed.
1. Only the signed out provider should be unregistered (we were
previously unregistering everything and reregistering again).

@:assistant