Fix intermittent issue applying mapped edits (#7949)
This change fixes an intermittent issue applying changes from a block in
Positron Assistant to the code editor.

We ask the model to generate a list of changes to apply and return the
list as JSON, which we then parse. However, the model sometimes returns
the JSON inside a Markdown code fence. That whole result gets parsed as
JSON and fails since code fences aren't JSON.

It's difficult to keep the model from doing this since most models seem
trained to do it. I have tried to work around this behavior as follows:
- Instructing the model more specifically not to use code fences in its
output.
- If the model _does_ return output wrapped in a JSON code fence,
discard the fence before parsing the JSON.

Addresses https://github.com/posit-dev/positron/issues/7188.

### QA Notes

This issue is very dependent on stochastic behavior from the model, and
does not reproduce reliably for me, so try it in several contexts.