Change settings import button to say "Compare settings" (#7990)
Addresses #7845 by changing the button text:

![Screenshot 2025-06-05 at 9 43
01 AM](https://github.com/user-attachments/assets/1a91d141-c43a-49cb-8827-ca370872eaae)

@:vscode-settings

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

The text on the button to import settings should be updated!