Update `RSTUDIO_PANDOC` with the correct location in Linux release builds (#7966)
This change fixes an issue that causes `RSTUDIO_PANDOC` to be empty in
Linux release builds

The issue is that on Linux, the layout of the Quarto distribution is
different than on other platforms; the tarball unpacks a top-level
folder `./quarto-1.2.3` containing all the files, so e.g. this is where
we are looking for Pandoc:

```
/resources/app/quarto/bin/tools/x86_64/pandoc
```

But it's actually here:

```
/resources/app/quarto/quarto-1.7.31/bin/tools/x86_64/pandoc
```

The fix is to remove the extra path segment after unpacking Pandoc on
Linux, so the layout is what we're expecting.

Addresses https://github.com/posit-dev/positron/issues/7929.

While I was in the files I made a few adjacent fixes:

- We now set `RSTUDIO_PANDOC` everywhere, not just in select tasks and R
sessions. Addresses https://github.com/posit-dev/positron/issues/7700.
- We also now set `QUARTO_R` in terminals. Addresses
https://github.com/posit-dev/positron/issues/7402 but not the larger
issue of setting up terminals for R which I didn't want to try to tangle
with this change.
- I've made the Environment Variables easier to configure; you can add
your own custom environment variables right in the settings UI, without
hand-editing JSON. Addresses
https://github.com/posit-dev/positron/issues/7875.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Custom environment variables for the Terminal and Consoles can now be
configured in the Settings UI (#7875)

#### Bug Fixes

- The `RSTUDIO_PANDOC` environment variable is now set correctly on
Linux builds (#7929)
- The `RSTUDIO_PANDOC` environment variable is now set in terminals and
tasks as well as R sessions (#7700)
- The `QUARTO_R` environment variable now points to your chosen R
version in the Terminal (#7402)


### QA Notes

Quarto (which includes Pandoc) gets added to the bundle during the
release build step, so local development builds will not have any Quarto
or Pandoc available, and it's expected that `RSTUDIO_PANDOC` is empty
for them. You'll want to use a release build to verify this change.