Data Explorer: Address DuckDB identifier quoting issues with fields containing non-ASCII characters  (#7972)
Addresses #7958. There were many places where a field identifier with
non-ASCII characters or quotations (double or single quotes) would
causes error's in DuckDB's SQL parser. This adds the appropriate quoting
and adds more robust logic in the generation of auxiliary field names
for the various statistics that are computed to power the summary pane.