python: keep asyncio logs from leaking into console (#8069)
I was working off this branch when doing package maintenance since a few
of the logs were popping up, and did not witness any other leaks 🎉

Unfortunately, this will hide logs if a user is 1) trying to use asyncio
in the console AND 2) using the default asyncio logger. Since a
StreamHandler or creating a logger allows folks to log as normal, I am
not super worried about this.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #7873 and #7995 Fixes bug where specific Python console logs were
printed to the user.


### QA Notes

Press F1 on a Python file/run some Python code. You should not see logs
like in #7873 (`SelectorSocketTransport` etc etc etc).