Send PID with heartbeat (address shutdown after wake from sleep) (#8048)
This change is another attempt to fix
https://github.com/posit-dev/positron/issues/6660.

After examining a fresh set of logs and doing some more testing on
Windows hardware, it appears that there's a problem with the idle
shutdown mechanism, which exists (in this case) as a failsafe to prevent
the supervisor from running forever after Positron exits. The mechanism
is incorrectly detecting that the supervisor is a zombie.

Recently, we added a client heartbeat endpoint to the supervisor which
keeps it alive as long as it receives regular pings from Positron.
However (you guessed it) Windows doesn't deliver these when the system
is asleep. So, on wake, the supervisor finds itself in a state where all
the kernels _and_ the frontend haven't been seen in hours (as all the
heartbeat mechanisms rely on network connections), and concludes that
it's become a zombie. It shuts itself down, ending sessions.
 
The fix is to send Positron's PID to the supervisor (with the
heartbeat), and update the supervisor so that it doesn't shut itself
down as long as that PID is still running. It also can bring the
supervisor back online if it exits unexpectedly.

Some other fixes/improvements:

- Explicitly forward `DYLD_LIBRARY_PATH` to kernel shells. Addresses
https://github.com/posit-dev/positron/issues/7859 and probably several
other issues related to R startup.
- Fix regression that caused many R startup problems to be reported as
handshake timeouts, and added new test to make sure kernel path exists
before trying to start it. Addresses
https://github.com/posit-dev/positron/issues/7941.
- Fix issue causing spurious heartbeat errors to be logged from the
supervisor during restart; these were the result of the heartbeat thread
not exiting quickly enough.
- Improve logging on the Positron side, with UTC timestamps that match
Kallichore's, so that the logs can be sorted to form correctly
interleaved output.
- Prevent multiple overlapping instances of "waiting for kernel to
reconnect/come back online" dialogs (something I noticed in a couple of
repro cases). Now we just show one.


### QA Notes

- I apologize in advance for the "leave laptop asleep overnight" part of
the repro.
- It is worth checking to be sure that Kallichore still gracefully
self-terminates after the idle timeout has expired, as most of this
change involves adding ways to prevent idle shutdown.
- If you set the shutdown timeout to "when idle", you'll get a 30 second
shutdown timer after Positron exits. This is the fastest idle timeout
available and may make it easier to test the change.