Assistant: Add Agent mode (#8098)
This is the minimal change to enable Agent mode in Positron Assistant.
There's more we'll want to do with this mode soon, but I'd like to get
it unlocked for folks.

Probably the most controversial part of this change is that I've enabled
pretty much all tools when using Agent mode (no filters). This gives
Agent access to things like MCP servers as well as a handful of built-in
tools and those provided by extensions, even if they are not tagged for
use with Assistant. Given the recent explosion in MCP servers and tool
availability I think we may need to reverse our approach and exclude a
list of known-problematic tools rather than requiring tools to opt in.

Addresses #8018. 

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Add Agent mode to Positron Assistant; note that you'll now need to use
this new mode if you want Assistant to run code on your behalf. (#8018)

#### Bug Fixes

- N/A


### QA Notes

As noted in the release notes, this change _removes_ the code execution
tool from _Ask_ and _Edit_ modes, meaning that Positron will only ever
suggest code in these modes.

Right now, if you try to configure the set of tools available to Agent,
you'll only see a couple of Jupyter tools (and ones that aren't relevant
at that). We will address this soon.

Test tags: `@:assistant`