Fix Copilot on Intel Macs (#8092)
This change fixes an issue that caused Copilot to fail on Intel Mac
computers.

The problem is that when we select the version of Copilot LSP binary to
bundle, we do so based on the `arch()` of the host system. Because all
our builds happen on arm64 hardware, however, this means that we only
bundle the arm64 binary.

The simplest fix would be to select the correct LSP binary based on
`npm_config_arch` in the x86_64 and arm64 build steps, and let the
universal app bundler create a single universal LSP binary from the two
copies. However, in my testing fusing the binary in this way made the
LSP stop working.

The approach taken here is to ship both the x86_64 *and* arm64 binaries
on macOS, and to select at runtime which one to use.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix error using Github Copilot with Intel Mac computers (#7969)


### QA Notes

I regret to inform you that it is worth checking to ensure Copilot is
still working on all 3 operating systems because the code affects where
the binary is placed and where we look for it afterwards.