Welcome Page Enhancements (#8118)
Addresses #3312 #7706 #3657 #7933

This PR implements the new welcome screen layouts. We've created two
separate experiences for users based off the workspace context. We want
to discourage users from doing work outside of a workspace so we now
show actions that lead users to a workspace on the welcome screen if
they aren't in a workspace. When a user is in a workspace we show them
file oriented actions.

The new layout also removes unnecessary actions (new console) and groups
like actions together. There are fewer overall actions shown which
should make it simple for users to get started quickly.

Design decisions:
- We wanted to maintain the large icon treatment for primary actions.
This layout does not support long labels so we have shortened "New
Folder from Template..." -> "New Folder..." and "New Folder from Git" ->
"New from Git...". These labels do not match what users see in dropdown.
We will see if users find this confusing, if they do, we can revisit the
layout to better support longer titles.
- We wanted to display the "Connect to..." action on the welcome page
but it is not important enough to be considered a primary action that
required the large icon treatment. We've tacked the action underneath
the recent folders list for now but we may want to revisit this design
if/when we add additional actions to be its own section. It does
currently feel a bit hard to see. We may need to revisit this based off
user feedback.
- The recent and help lists have swapped locations since we want to keep
similar actions grouped together. This layout ensures the recent list is
always visible with primary actions on smaller window sizes now.
- The help section now resides next to the primary actions view instead
of next to the recently opened list view. This is intentional since we
want to add walkthroughs back to the welcome screen. The work to align
the help and recent sections would need to be undone once walkthroughs
are added.

### Screenshots

<img width="306" alt="Screenshot 2025-06-13 at 5 27 17 PM"
src="https://github.com/user-attachments/assets/113d1e09-3ca3-4372-9681-3b5287265e8f"
/>


https://github.com/user-attachments/assets/345e480e-5f8f-4f66-90fb-48697e785983


https://github.com/user-attachments/assets/f18cfb69-685f-4880-ab24-415aaa367045

### Release Notes

#### New Features

- Customize welcome page screen for workspace and non-workspace context
- Display company name in About dialog

#### Bug Fixes

- Fix incorrect hover label for "New Folder from Git" action

### QA Notes

@:web @:win @:welcome