Assistant: Improve default prompt (#8106)
Work towards:
 * #8128 
 
Addresses:
 * #8050 
 * #7821.

---

* Rewrites the default prompt to better steer how we want the Positron
Assistant to behave.

* Includes some data science examples in R and Python for the assistant
to work from.

* Adds the current date and time, in the host's current locale, in the
Positron context. (c.f. https://github.com/posit-dev/positron/pull/8057)

### QA Notes

Due to the stochastic nature of LLM responses, I don't think this will
be easy to test until we have an LLM evaluation framework. The next best
thing would probably be to chat with the assistant for a little while
and see how it feels.

For the date/time context, simply ask the bot for the current date or
time.