Fix frozen console after extension host restart (#8137)
This change addresses a few problems that have accumulated in the
"restart extension host" scenario.

- The console is frozen. This is a regression from
https://github.com/posit-dev/positron/pull/6832; which returns early
before detaching the session; because we never detached the existing
runtime session, the new session was not reattached and console stayed
attached to the old (disconnected) session.
- It is no longer possible to open data explorers from the Variables
pane. As far as I can tell this never worked; fixed by reattaching event
handlers to the new session.
- The logs contain errors about registered runtimes existing on the main
thread but not the extension host, which happens because restarting the
extension host causes it to forget its runtimes (of course). After
auditing the code it appears that it isn't actually necessary for the
extension host to keep a copy of the registration list; fixed by having
this knowledge exist only on the main thread, to be queried by the
extension host when required.
- Various errors about leaked disposables; addressed by making the
session adapter disposable and making sure event handlers get registered
for disposal.

I also found a couple of places where exceptions were not propagated,
which is why the console looks frozen instead of popping up an error. I
added notifications to those places, so if the bug were to regress the
user would get an error instead of nothing at all.

We should also at some point think about whether there is a better way
to manage `ILanguageRuntimeSession` objects. Right now an extension host
restart results in a new `ILanguageRuntimeSession` object with the same
session ID as the previous one; it is imperative that everyone working
with the old one switch to the new one, but this requires verbose and
error-prone code. We could make this better (but not before RC).

Addresses https://github.com/posit-dev/positron/issues/8096

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix frozen console after changing an extension
(https://github.com/posit-dev/positron/issues/8096)


### QA Notes

It is not necessary to go through all the steps in the user-filed issue
above to trigger this problem; there is a _Restart Extension Host_
command that forces an ext host restart. This will make the problem
happen every time.

<img width="620" alt="image"
src="https://github.com/user-attachments/assets/b9246856-99d9-419d-8fe4-08213197225b"
/>

This PR doesn't fix every problem with extension host restarts; if you
notice an issue, check to see if it repros on main.

The affected code is also involved in the codepath that runs when you
switch projects in the same window, an area that's been regression
prone, so spot test that too.

@testlabauto offered to write a test that goes through the extension
host restart process, but I'd also be glad to write one if needed.

Test tags: `@:console`