Assistant: Improve completions when using FIM prompting (#8030)
Improves the experience when using a FIM prompt for ghost text
completion.

Note: This method is _not_ used for GitHub Copilot completions, but can
enable ghost text when using API providers without built-in completion
support.

Changes:

* Rewritten and improved the FIM completion prompt.
* Switched from XML to using the "standard" FIM completion tokens:
`<|fim_prefix|>`, `<|fim_suffix|>`, `<|fim_middle|>`. IIUC many modern
LLM models now support these tokens natively.
* Fixed including related content from additional files in the workspace
in the FIM request.
* Lowered temperature (following OpenAI documentation) and added stop
sequences.
* Updated some LLM model selection:
  - OpenAI: GPT-4.1 Mini (for completion only)
  - Google: Gemini 2.0 Flash (for completion only)
  - Mistral: Mistral Medium (chat), Codestral (completion)
 
* Added a parameter `completion` to provider config. When `true`, signs
in to both a chat and completion source when logging in with that
provider.
* Set the above `completion` parameter to `true` for Mistral, Google,
and OpenAI.

### QA Notes

It is intended that the "default available" providers Anthropic and
GitHub Copilot are unaffected by this change.

To test:
* Add provider to experimental `positron.assistant.enabledProviders`
setting.
* Login with a Mistral, Google or OpenAI API key
* Observe that a completion model is registered alongside the chat model
* Confirm ghost text completions appear from the newly registered model